+++ go.mod
@@ -28,7 +28,6 @@ require (
 
 require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/mitchellh/mapstructure v1.5.0
 	github.com/spf13/pflag v1.0.5 // indirect
 	golang.org/x/tools v0.15.0
 )
